new syntax
git-svn-id: svn://localhost/trunk@1031 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/pnts_internal.cpp
@@ -323,7 +323,7 @@ void prepare_scattered_points(const d_points * pnts,
 	
 	std::vector<size_t> * point_numbers = new std::vector<size_t>( pnts->size() );
 	
-	int point_counter;
+	size_t point_counter;
 	for (point_counter = 0; point_counter < pnts->size(); point_counter++) {
 		*( point_numbers->begin() + point_counter ) = point_counter;
 	}
